+++ lib/auth.js
@@ -1,6 +1,6 @@
 import Router from 'next/router';
 import Cookies from 'js-cookie';
-import { fetcher } from '../components/tutorial/NavAuth';
+import { fetcher } from './api';
 
 export const setToken = (data) => {
   if (typeof window === 'undefined') {
